Do not unnecessarily rewrite output on update
Prior to this change, the modification timestamp of all template output
is updated every time buildout is run, even when the content of the
output has not changed.  This changes things so that the output file is
rewritten (and its mtime updated) on update only if its content has
changed.

This can cause unnecessary grief in some circumstances, depending on the
consure of the template output. As an example, systemd checks the
timestamps on its unit files, and, if any have changed, issues a helpful
"please run systemctl daemon-reload" warning messages.